Move selectors to their corresponding modules
