+++ src/layouts/show-dish.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { removeDish} from '../modules/dishes'
 import { Link } from 'react-router'
-import { totalSelector, escandalloSelector } from '../selectors/dishSelectors'
+import { totalSelector, escandalloSelector } from '../modules/dishes/selectors'
 
 class ShowDish extends Component {
   onRemove() {
